refactor: move Repository Interface to Domain
