+++ app/src/main/java/campus/tech/kakao/map/domain/usecase/GetAllSearchWordsUseCase.kt
@@ -1,7 +1,6 @@
 package campus.tech.kakao.map.domain.usecase
 
-import campus.tech.kakao.map.data.model.SavedSearchWord
-import campus.tech.kakao.map.data.repository.SavedSearchWordRepository
+import campus.tech.kakao.map.domain.repository.SavedSearchWordRepository
 import campus.tech.kakao.map.domain.model.SavedSearchWordDomain
 import javax.inject.Inject
