+++ app/src/main/java/campus/tech/kakao/map/domain/usecase/InsertOrUpdateSearchWordUseCase.kt
@@ -1,6 +1,6 @@
 package campus.tech.kakao.map.domain.usecase
 
-import campus.tech.kakao.map.data.repository.SavedSearchWordRepository
+import campus.tech.kakao.map.domain.repository.SavedSearchWordRepository
 import campus.tech.kakao.map.domain.model.SavedSearchWordDomain
 import javax.inject.Inject
